probar

+++ Funcionalidades/Persona/persona.cpp
@@ -19,11 +19,10 @@ void Persona::CambiarNombre(const char *NuevoNombre)
 	strncpy(this->nombre, NuevoNombre, 49);
 	this->nombre[49] = '\0';
 }
-void Persona::Cambiar_Id(size_t nuevo_id){
+void Persona::Cambiar_Id(int nuevo_id){
 	cin>>nuevo_id;
 	this->id= nuevo_id;
 }
-
 
 vector<Libro> Persona::VerLibros(string nombreArchivo,bool crear){
 	ifstream archi(nombreArchivo,ios::binary);
